增加新闻界面返回

+++ jteamAction.php
@@ -14,10 +14,19 @@ if (!empty($_GET['member_sno'])&&!empty($_GET['cap_sno'])&&$member_sno!=$cap_sno
         $conn->query($sql);
         header("location:jteam.php?id=".$id);
     }else{
-        $sql_c = "delete from static where membersno = $member_sno and capsno = $cap_sno";
-        $conn->query($sql_c);
-        echo "<script>var d = \"$id\"</script>";
-        echo "<script>alert('取消成功！');window.location.href='jteam.php?id='+d</script>";
+        $sql_cen = "select static_join from static where membersno = $member_sno and capsno = $cap_sno";
+        $result_cen = $conn->query($sql_cen);
+        $row_cen = mysqli_fetch_array($result_cen);
+        if ($row_cen['static_join'] == 2||$row_cen['static_join'] == 3){
+            echo "<script>var d = \"$id\"</script>";
+            echo "<script>alert('操作失败，队长已处理！');window.location.href='jteam.php?id='+d</script>";
+        }else{
+            $sql_c = "delete from static where membersno = $member_sno and capsno = $cap_sno";
+            $conn->query($sql_c);
+            echo "<script>var d = \"$id\"</script>";
+            echo "<script>alert('取消成功！');window.location.href='jteam.php?id='+d</script>";
+        }
+
     }
 }else{
     echo "<script>var d = \"$id\"</script>";
